requirements.txt
beautifulsoup4
lxml
requests
pandas
pyyaml
rapidfuzz
selenium
webdriver-manager
scrapy